backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
email-validator==2.2.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
pymilvus==2.6.0
sentence-transformers==3.3.1
pdfplumber==0.11.5
python-docx==1.1.2
tenacity==9.0.0
langfuse==3.10.5
prometheus-fastapi-instrumentator==7.0.0
